Frontend/requirements.txt
### Requirements for front end

pyserial
streamlit
